example: demonstrate tracepoint usage
