+++ .semaphore/semaphore.yml
@@ -31,6 +31,7 @@ blocks:
         commands:
           - cache store
           - git diff --exit-code || { echo "generated files are not up to date" >&2; false; }
+          - for i in $(find _example -maxdepth 1 -mindepth 1 -type d); do cd $i; go build; cd -; done
       env_vars:
         - name: TMPDIR
           value: /tmp
