+++ _example/tracepoint/go.mod
@@ -1,0 +1,9 @@
+module github.com/cilium/ebpf/_example/tracepoint
+
+go 1.15
+
+require (
+	github.com/cilium/ebpf v0.3.0
+	github.com/elastic/go-perf v0.0.0-20191212140718-9c656876f595
+	golang.org/x/sys v0.0.0-20200124204421-9fbb57f87de9
+)
